Add `PriceCalculator` for discounted pricing and update `OrderService`
Introduce `PriceCalculator` to handle discounted pricing logic, update `OrderService` to use it, and revise schema to store product prices as `DECIMAL(5, 2)`. Add corresponding tests for `PriceCalculator`.

+++ src/main/java/ch/martinelli/demo/nomocks/query/Product.java
@@ -1,4 +1,6 @@
 package ch.martinelli.demo.nomocks.query;
 
-record Product(Long id, String name, double price) {
+import java.math.BigDecimal;
+
+record Product(Long id, String name, BigDecimal price) {
 }
